Option to remove Median Warm-up temp
I personally disagree with "delta T over ambient" temperatures because they hide useful information, but I can see a use to it
There is a checkbox toggle now for subtracting the median Warm-up temperature, a proxy for ambient temperature, that will subtract that amount from the appropriate values in the appropriate tables.
I believe this median can be considered representative of the ambient temperature, especially as it is sensitive to the same environmental variables as the load temperatures. It has to be sensitive as it's the same measurement, but at a different time.