Correction to Temperature Offset
For the temperature offset feature, it is necessary to subtract the value from the input, not the output (for tableECDF, not tablePERC but I want these functions to be similar). This makes those functions a little messier, but the result will be better.
It should be noted though that the target value for ECDF does not change with the subtraction, but the data it is applied to does.