Histogram Differences improvements
The difference controls on the Histogram pages cannot allow values of 0, so the current bin width is forced.
Bin width's cannot be zero, so the defaults are set if 0 is selected. (Setting a minimum of 0 for the field is not ideal as it offsets the up and down controls.)
Values in the limit fields will flip if the maximum is set below the minimum.
Identical values for min and max will cause the table to not be made.