FREQ.COEF changes
Apply Coefficient button for safer editing
Decided to invert the FREQ.COEF value so it is more clear. Rather than being the coefficient on the frequency, it will be the coefficient on the power, which will be an integer as opposed to a fraction.
It also makes more sense with the name "Power-Frequency Coefficient".