Fix for brush bouncing
I'm hopefully that this will address the issue of the brush limits bouncing between values. This solution is to properly abstract the values to brushHISTxxxx so that the input update is only tied to that. Previously a change to the input could also trigger the input to update and that I believe was causing the bouncing. Now it is only a change to the abstracted variable that will update the input.
Technically this does mean using the input will also trigger a change here, but that should not be a problem.