Improved Histogram Modality Modularity v1
While this is technically a better modular implementation, it also will not clear the table when new data is loaded, but instead recalculates the table.
I cannot find any solution to this issue because, for some reason, the moduleServer function is ignorant of anything from the host namespace and I cannot pass the information it to effectively.