Improved Histogram Modality Modularity v2
Found a solution to the issue by exploiting something I noticed in an error. It is possible to explicitly call a value from a namespace outside a modular function by appending the proper prefix.
This way I can call the update button in a different context where the other buttons are accessible.
Technically this is still cleaner than before, as the namespaces will keep most objects separated, but now there must be unique calls to the update buttons and unique IDs for the tables, so this is not as modular as I wish.
It is that separation of objects by a name space that makes a standard custom function not so desirable here.