Improved Time Series Modularity
For most of this, the modularity is somewhat better, with the suggested means of modularizing Shiny code
The button to apply the zoom to other season graphs, however, is effectively the same as before, but now it uses moduleServer and namespaces. I am piercing into namespaces to get specific values often, so the modular benefits are limited. Really, it's just about getting namespaces working with it.