Correcting NS object names
In the article on modularizing Shiny code, it consistently uses "id" as the object name. The thing is, the NS function has an argument named "id" but that is not the one being used here. Instead it is the "namespace" argument being used, so for clarity I have changed to using "name" as the object name.